% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_info.R
\name{cache_info}
\alias{cache_info}
\title{List folders and their sizes in cache}
\usage{
cache_info(dir = NULL)
}
\arguments{
\item{dir}{Character, path to directory where data will be stored. If
\code{NULL}, \link[rappdirs]{user_cache_dir} will be used to determine path.}
}
\value{
A tibble with 2 variables:
\describe{
  \item{name}{Name of the folder}
  \item{size}{Size of the folder}
}
}
\description{
This function will return a tibble with the name and sizes of all folder in
specified directory. Will default to textdata's default cache.
}
\examples{
\dontrun{
cache_info()
}
}
