% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afinn.R
\name{dataset_afinn}
\alias{dataset_afinn}
\title{AFINN-111 dataset}
\usage{
dataset_afinn()
}
\value{
A tibble with 2 columns.
}
\description{
AFINN is a list of English words rated for valence with an integer
between minus five (negative) and plus five (positive). The words have
been manually labeled by Finn Årup Nielsen in 2009-2011.
}
\details{
This dataset is the newest version with 2477 words and phrases.
}
