% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_bing.R
\name{dataset_bing}
\alias{dataset_bing}
\title{Bing Sentiment Lexicon}
\usage{
dataset_bing(...)
}
\arguments{
\item{...}{Additional arguments passed to [load_dataset()].}
}
\value{
A tibble with 2 columns.
}
\description{
Bing Sentiment Lexicon
}
