% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loughran.R
\name{dataset_loughran}
\alias{dataset_loughran}
\title{Loughran-McDonald Sentiment Word Lists}
\usage{
dataset_loughran(...)
}
\arguments{
\item{...}{Additional arguments passed to [load_dataset()].}
}
\value{
A tibble with 2 columns.
}
\description{
Loughran-McDonald Sentiment Word Lists
}
