% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_sentence_polarity.R
\name{dataset_sentence_polarity}
\alias{dataset_sentence_polarity}
\title{v1.0 sentence polarity dataset}
\source{
\url{https://www.cs.cornell.edu/people/pabo/movie-review-data/}
}
\usage{
dataset_sentence_polarity(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{load_dataset}}.}
}
\value{
A tibble with 10,662 rows and 2 variables:
\describe{
  \item{text}{Sentences or snippets}
  \item{sentiment}{Indicator for sentiment, "neg" for negative and "pos"
                   for positive}
}
}
\description{
5331 positive and 5331 negative processed sentences / snippets.
Introduced in Pang/Lee ACL 2005. Released July 2005.
}
\details{
Citation info:

This data was first used in Bo Pang and Lillian Lee,
``Seeing stars: Exploiting class relationships for sentiment categorization
with respect to rating scales.'', Proceedings of the ACL, 2005.

InProceedings\{pang05, \cr
author    = \{Bo Pang and Lillian Lee\}, \cr
title     = \{Seeing stars: Exploiting class relationships for sentiment \cr
              categorization with respect to rating scales\}, \cr
booktitle = \{Proceedings of the ACL\}, \cr
year      = 2005 \cr
\}
}
\keyword{datasets}
