% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_functions.R
\docType{data}
\name{download_functions}
\alias{download_functions}
\title{List of all download functions used in load_dataset}
\format{Named list of all download functions}
\usage{
download_functions
}
\description{
List of all download functions used in load_dataset
}
\keyword{datasets}
