% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_afinn.R
\name{lexicon_afinn}
\alias{lexicon_afinn}
\title{AFINN-111 dataset}
\usage{
lexicon_afinn(...)
}
\arguments{
\item{...}{Additional arguments passed to [load_dataset()].}
}
\value{
A tibble with 2 columns.
}
\description{
AFINN is a list of English words rated for valence with an integer
between minus five (negative) and plus five (positive). The words have
been manually labeled by Finn Årup Nielsen in 2009-2011.
}
\details{
This dataset is the newest version with 2477 words and phrases.
}
