% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_bing.R
\name{lexicon_bing}
\alias{lexicon_bing}
\title{Bing sentiment lexicon}
\source{
\url{https://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\usage{
lexicon_bing(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{load_dataset}}.}
}
\value{
A tibble with 6,787 rows and 2 variables:
\describe{
  \item{word}{An English word}
  \item{sentiment}{Indicator for sentiment: "negative" or "positive"}
}
}
\description{
General purpose English sentiment lexicon that categorizes words in a
binary fashion, either positive or negative
}
\keyword{datasets}
