% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_loughran.R
\name{lexicon_loughran}
\alias{lexicon_loughran}
\title{Loughran-McDonald Sentiment Word Lists}
\usage{
lexicon_loughran(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{load_dataset}}.}
}
\value{
A tibble with 2 columns.
}
\description{
Loughran-McDonald Sentiment Word Lists
}
