% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\docType{data}
\name{process_functions}
\alias{process_functions}
\title{List of all process functions used in load_dataset}
\format{Named list of all process functions}
\usage{
process_functions
}
\description{
List of all process functions used in load_dataset
}
\keyword{datasets}
