% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textdata-package.R
\docType{package}
\name{textdata-package}
\alias{textdata}
\alias{textdata-package}
\title{textdata: Download and Load Various Text Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a framework to download, parse, and store text datasets on the disk and load them when needed. Includes various sentiment lexicons and labeled text data sets for classification and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilhvitfeldt.github.io/textdata/}
  \item \url{https://github.com/EmilHvitfeldt/textdata}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/textdata/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Other contributors:
\itemize{
  \item Julia Silge \email{julia.silge@gmail.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID}) [contributor]
}

}
\keyword{internal}
