% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{catalogue}
\alias{catalogue}
\title{Catalogue of all available data sources}
\format{
An object of class \code{data.frame} with 15 rows and 8 columns.
}
\usage{
catalogue
}
\description{
Catalogue of all available data sources
}
\keyword{datasets}
