% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_polarity.R
\name{dataset_sentence_polarity}
\alias{dataset_sentence_polarity}
\title{v1.0 sentence polarity dataset}
\usage{
dataset_sentence_polarity()
}
\value{
A tibble with 2 columns.
}
\description{
5331 positive and 5331 negative processed sentences / snippets.
Introduced in Pang/Lee ACL 2005. Released July 2005.
}
