% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_afinn.R
\name{lexicon_afinn}
\alias{lexicon_afinn}
\title{AFINN-111 dataset}
\usage{
lexicon_afinn(dir = NULL, delete = FALSE, return_path = FALSE)
}
\arguments{
\item{dir}{Character, path to directory where data will be stored. If
\code{NULL}, \link[rappdirs]{user_cache_dir} will be used to determine path.}

\item{delete}{Logical, set \code{TRUE} to delete dataset.}

\item{return_path}{Logical, set \code{TRUE} to return the path of the dataset.}
}
\value{
A tibble with 2,477 rows and 2 variables:
\describe{
  \item{word}{An English word}
  \item{sentiment}{Indicator for sentiment: integer between -5 and +5}
}
}
\description{
AFINN is a lexicon of English words rated for valence with an integer
between minus five (negative) and plus five (positive). The words have
been manually labeled by Finn Årup Nielsen in 2009-2011.
}
\details{
This dataset is the newest version with 2477 words and phrases.

Citation info:

This dataset was published in Finn Ärup Nielsen (2011),
``A new {ANEW:} Evaluation of a word list for sentiment analysis in
microblogs'', Proceedings of the ESWC2011 Workshop on
'Making Sense of Microposts': Big things come in small packages (2011) 93-98.

article\{nielsen11, \cr
author    = \{Finn Äruprup Nielsen\}, \cr
title     = \{A new {ANEW:} Evaluation of a word list for sentiment analysis in microblogs\}, \cr
journal   = \{CoRR\}, \cr
volume    = \{abs/1103.2903\}, \cr
year      = \{2011\}, \cr
url       = \{http://arxiv.org/abs/1103.2903\}, \cr
archivePrefix = \{arXiv\}, \cr
eprint    = \{1103.2903\}, \cr
biburl    = \{https://dblp.org/rec/bib/journals/corr/abs-1103-2903\}, \cr
bibsource = \{dblp computer science bibliography, https://dblp.org\} \cr
\}
}
\examples{
\donttest{
lexicon_afinn()

# Custom directory
lexicon_afinn(dir = "data/")

# Deleting dataset
lexicon_afinn(delete = TRUE)

# Returning filepath of data
lexicon_afinn(return_path = TRUE)
}
}
\seealso{
Other lexicon: \code{\link{lexicon_bing}},
  \code{\link{lexicon_loughran}}, \code{\link{lexicon_nrc}}
}
\concept{lexicon}
\keyword{datasets}
