% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_bing.R
\name{lexicon_bing}
\alias{lexicon_bing}
\title{Bing Sentiment Lexicon}
\usage{
lexicon_bing(...)
}
\arguments{
\item{...}{Additional arguments passed to [load_dataset()].}
}
\value{
A tibble with 2 columns.
}
\description{
Bing Sentiment Lexicon
}
