% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon_loughran.R
\name{lexicon_loughran}
\alias{lexicon_loughran}
\title{Loughran-McDonald sentiment lexicon}
\source{
\url{https://sraf.nd.edu/textual-analysis/resources/}
}
\usage{
lexicon_loughran(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{load_dataset}}.}
}
\value{
A tibble with 4,150 rows and 2 variables:
\describe{
  \item{word}{An English word}
  \item{sentiment}{Indicator for sentiment: "negative", "positive",
  "litigious", "uncertainty", "constraining", or "superfluous"}
}
}
\description{
English sentiment lexicon created for use with financial documents. This
lexicon labels words with six possible sentiments important in financial
contexts: "negative", "positive", "litigious", "uncertainty", "constraining",
or "superfluous".
}
\details{
Citation info:

This dataset was published in Loughran, T. and McDonald, B. (2011),
``When Is a Liability Not a Liability? Textual Analysis, Dictionaries, and
10‐Ks.'' The Journal of Finance, 66: 35-65.

article\{loughran11, \cr
author  = \{Loughran, Tim and McDonald, Bill\}, \cr
title   = \{When Is a Liability Not a Liability? Textual Analysis, Dictionaries, and 10-Ks\}, \cr
journal = \{The Journal of Finance\}, \cr
volume  = \{66\}, \cr
number  = \{1\}, \cr
pages   = \{35-65\}, \cr
doi     = \{10.1111/j.1540-6261.2010.01625.x\}, \cr
url     = \{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1540-6261.2010.01625.x\}, \cr
eprint  = \{https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1540-6261.2010.01625.x\}, \cr
year    = \{2011\} \cr
\}
}
\keyword{datasets}
