% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Internal Functions}
\usage{
load_dataset(data_name, name, dir, delete, return_path, clean,
  clean_manual = NULL)
}
\description{
These are not to be used directly by the users.
}
\keyword{internal}
